Fix incorrect font usage in death screen
This commit corrects the usage of a wrong font object in the game's death screen, changing 'fontsbig' to the correct 'fontbig'. This was causing a potential error where the associated text could not be rendered properly. With this adjustment, the deathmessage text will now be displayed correctly enhancing the user experience.