Adjust main.py to allow custom tilemap sizes
Updated the create_tiles function in the main.py script to allow for customization of the tilemap size. Previously, tilemap was fixed at a size of 100x100. This change allows for variable map sizes to be specified as parameters when creating tiles. The modification, whilst preserving the previous default behavior, contributes to increased flexibility when generating game scenarios or testing different map configurations.