Increase mapsize in PerlinNoise function
Updated the line in which the PerlinNoise function was called by adding a 'mapsize' parameter. This allows for the creation of a larger tilemap resulting in a more diverse game environment. The 'mapsize' parameter is set to 200 which doubles the previous default size. This would accommodate larger game scenarios and enable broader testing capabilities.