Improve PyInstaller settings for smoother software distribution
This commit introduces changes on `main.py` and `main.spec` to enhance and streamline our deployment process using PyInstaller. Now, when running a bundled app, we change the working directory to the directory where the bundled app lives. This ensures file paths are calculated correctly. In addition, `main.spec` file is updated to include 'sounds' in distribution packaging. Also updated '.gitignore' to prevent accidental push of binary files to the repository.